fx: move Stairs2Slope
