+++ meson.build
@@ -103,6 +103,7 @@ sources = [
   'src/game/effect_routines/powerup.c',
   'src/game/effect_routines/raising_block.c',
   'src/game/effect_routines/sand.c',
+  'src/game/effect_routines/stairs2slope.c',
   'src/game/effects/blood.c',
   'src/game/effects/exploding_death.c',
   'src/game/effects/gun.c',
@@ -165,7 +166,6 @@ sources = [
   'src/game/objects/effects/missile.c',
   'src/game/objects/effects/ricochet.c',
   'src/game/objects/effects/splash.c',
-  'src/game/objects/effects/stairs2slope.c',
   'src/game/objects/effects/turn_180.c',
   'src/game/objects/effects/twinkle.c',
   'src/game/objects/effects/waterfall.c',
