+++ src/game/effect_routines/stairs2slope.c
@@ -1,9 +1,9 @@
-#include "game/objects/effects/stairs2slope.h"
+#include "game/effect_routines/stairs2slope.h"
 
 #include "game/sound.h"
 #include "global/vars.h"
 
-void Stairs2Slope(ITEM_INFO *item)
+void FX_Stairs2Slope(ITEM_INFO *item)
 {
     if (g_FlipTimer == 5) {
         Sound_Effect(SFX_STAIRS2SLOPE_FX, NULL, SPM_NORMAL);
